Change on Inversionist and Inverst

+++ Backend/src/main/java/com/example/starfund/domain/repository/InversionRepository.java
@@ -5,7 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface InversionRepository extends JpaRepository<Inversion, Integer> {
+public interface InversionRepository extends JpaRepository<Inversion, Long> {
 
-   //void findByInversionistaId(int inversionista_id);
 }
