+++ Backend/src/main/java/com/example/starfund/domain/repository/InversionistaRepository.java
@@ -1,10 +1,16 @@
 package com.example.starfund.domain.repository;
 
+
 import com.example.starfund.domain.model.entity.Inversionista;
+
+
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface InversionistaRepository extends JpaRepository<Inversionista, Integer> {
+public interface InversionistaRepository extends JpaRepository<Inversionista, Long> {
+
+    
 
 }
