+++ Backend/src/main/java/com/example/starfund/domain/repository/StartupRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface StartupRepository extends JpaRepository<Startup, Integer> {
+public interface StartupRepository extends JpaRepository<Startup, Long> {
 
 
 }
